+++ src/Sulu/Bundle/ContactBundle/Resources/config/services.xml
@@ -64,6 +64,7 @@
         <service id="sulu_contact.account_manager" class="%sulu_contact.account_manager.class%">
             <argument type="service" id="doctrine.orm.entity_manager"/>
             <argument type="service" id="sulu_tag.tag_manager"/>
+            <argument type="service" id="sulu_contact.account_factory"/>
             <argument type="string">%sulu_contact.account.entity%</argument>
         </service>
         <service id="sulu_contact.contact_manager" class="%sulu_contact.contact_manager.class%">
