+++ src/Sulu/Bundle/ContactBundle/Resources/public/js/components/accounts/components/form/main.js
@@ -41,7 +41,7 @@ define(['config', 'widget-groups'], function(Config, WidgetGroups) {
         layout: function() {
             return {
                 content: {
-                    width: (WidgetGroups.exists('account-detail') ? 'max' : 'fixed')
+                    width: (WidgetGroups.exists('account-detail') && !!this.options.data && !!this.options.data.id ? 'max' : 'fixed')
                 },
                 sidebar: {
                     width: 'fixed',
